09/11/2023 check

+++ STM32/PROCESSING/SENSO/senso.h
@@ -20,7 +20,7 @@ typedef struct level
     volatile uint32_t SR;
     uint16_t sizeResponse;
 
-    float levelValue;
+    uint8_t levelValue;
     uint8_t angle;
     uint8_t valueIndex;
     uint32_t install;
@@ -31,7 +31,7 @@ typedef struct level
     CTL_StatusTypedef (*levenGetConfig)(struct level *const me, uint32_t configType, char pData[], uint16_t size);
     CTL_StatusTypedef (*levsenSetSection)(struct level *const me, uint8_t section);
 
-    float (*levenGetLevel)(struct level *const me);
+    uint8_t (*levenGetLevel)(struct level *const me);
 
 } LEVEL_HandleTypedef;
 
@@ -40,7 +40,7 @@ typedef CTL_StatusTypedef (*levenDeInit)(struct level *const me);
 typedef CTL_StatusTypedef (*levenProcess)(struct level *const me, uint16_t size);
 typedef CTL_StatusTypedef (*levenGetConfig)(struct level *const me, uint32_t configType, char pData[], uint16_t size);
 typedef CTL_StatusTypedef (*levsenSetSection)(struct level *const me, uint8_t section);
-typedef float (*levenGetLevel)(struct level *const me);
+typedef uint8_t (*levenGetLevel)(struct level *const me);
 
 #define SENSO_FLAG_SCF 0x01 // SENSO CONNECT FAIL
 #define SENSO_FLAG_RX 0x02
@@ -68,7 +68,7 @@ CTL_StatusTypedef SENSOR_deInit(LEVEL_HandleTypedef *const me);
 CTL_StatusTypedef SENSOR_process(LEVEL_HandleTypedef *const me, uint16_t size);
 CTL_StatusTypedef SENSOR_getConfig(LEVEL_HandleTypedef *const me, uint32_t configType, char pData[], uint16_t size);
 CTL_StatusTypedef SENSOR_setSection(LEVEL_HandleTypedef *const me, uint8_t section);
-float SENSOR_getNewValue(LEVEL_HandleTypedef *const me);
+uint8_t SENSOR_getNewValue(LEVEL_HandleTypedef *const me);
 
 CTL_StatusTypedef SENSOR_setAngle(LEVEL_HandleTypedef *const me, uint8_t value);
 CTL_StatusTypedef SENSO_setInstall(LEVEL_HandleTypedef *const me, uint8_t value);
